resolve pylint no eps member error

+++ bambi/families/link.py
@@ -9,7 +9,7 @@ from bambi.utils import multilinify, indentify
 
 def force_within_unit_interval(x):
     """Make sure data in unit interval is in (0, 1)."""
-    eps = np.finfo(float).eps
+    eps = np.finfo(np.float64).eps
     x[x == 0] = eps
     x[x == 1] = 1 - eps
     return x
@@ -17,7 +17,7 @@ def force_within_unit_interval(x):
 
 def force_greater_than_zero(x):
     """Make sure data in positive reals is in (0, infty)."""
-    eps = np.finfo(float).eps
+    eps = np.finfo(np.float64).eps
     x[x == 0] = eps
     return x
